% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVOLib.R
\name{vvo_getDeps}
\alias{vvo_getDeps}
\title{Get Departures by StopID}
\usage{
vvo_getDeps(stopid, limit = 10)
}
\arguments{
\item{stopid}{StopID (can be obtained by vvo_getStops)}

\item{limit}{Resultcount. Default 10, Maximum (due to API-Restrictions) 50}
}
\value{
Tibble with Informations about the Departures
}
\description{
Get Departures by StopID
}
