% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVOLib.R
\name{vvo_getPuR}
\alias{vvo_getPuR}
\title{Get Park & Ride Places in VVO}
\source{
http://www.vvo-online.de/open_data/PuR.JSON
}
\usage{
vvo_getPuR()
}
\value{
Tibble with all P&R Places in VVO, fetched from an open_data-JSON. There might be duplicates.
}
\description{
Get Park & Ride Places in VVO
}
