% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVOLib.R
\name{vvo_getRouteChanges}
\alias{vvo_getRouteChanges}
\title{Get all RouteChanges in VVO}
\usage{
vvo_getRouteChanges()
}
\value{
Tibble with all Changes in the VVO.
}
\description{
Get all RouteChanges in VVO
}
