% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVOLib.R
\name{vvo_getStops}
\alias{vvo_getStops}
\title{Get Stops in VVO based on a query}
\usage{
vvo_getStops(query)
}
\arguments{
\item{query}{Query to search with. Supports Abbrevations by DVB}
}
\value{
Tibble with Informations about the Stop
}
\description{
Get Stops in VVO based on a query
}
\examples{
> vvo_getStops("LAI")
# A tibble: 1 x 5
ID City    Station          GK_Right   GK_Up
<int> <chr>   <chr>               <int>   <int>
1 33000076 Dresden Laibacher Straße  5655656 4628750
}
