% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVOLib.R
\name{vvo_getTrip}
\alias{vvo_getTrip}
\alias{vvo_getTripByDep}
\title{Get a Trip by StopID, TripID and Time}
\usage{
vvo_getTrip(stopid, tripid, time = Sys.time())

vvo_getTripByDep(stopid, departure)
}
\arguments{
\item{stopid}{StopID (can be obtained by vvo_getStops)}

\item{tripid}{TripID (can be obtained by vvo_getDeps). Note that its not an ID for a specific trip, but for a route (e.g. all Trips on line 1 from Prohlis to Leutewitz have the same ID)}

\item{time}{Departuretime of the specific trip on the specified station. Default is current time.}

\item{departure}{One observation from vvo_getDeps}
}
\value{
Tibble with Informations about all Stations in the Trip
}
\description{
Get a Trip by StopID, TripID and Time

... or by StopID and Departure
}
