% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVOLib.R
\name{vvo_transformPlatformType}
\alias{vvo_transformPlatformType}
\title{Helper-Method to beautify Platform-Description}
\usage{
vvo_transformPlatformType(Platform)
}
\arguments{
\item{Platform}{Platformobject from other Functions}
}
\value{
String with Steig or Gleis depending on Platformtype
}
\description{
Helper-Method to beautify Platform-Description
}
