% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVOLib.R
\name{vvo_getLinesAtStop}
\alias{vvo_getLinesAtStop}
\title{Get Lines at Stop by StopID}
\usage{
vvo_getLinesAtStop(stopid)
}
\arguments{
\item{stopid}{StopID (can be obtained by vvo_getStops)}
}
\value{
Tibble with all Lines at this stop.
}
\description{
Get Lines at Stop by StopID
}
